requirements.txt
streamlit
torch
torchvision
numpy
matplotlib
opencv-python
Pillow
